# docs/swagger.json
{
    "openapi": "3.1.0",
    "info": {
      "version": "1.0.0",
      "title": "Contacts manager",
      "license": {
        "name": "Apache 2.0",
        "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
      },
      "description": "contact management."
    },
    "servers": [
      {
        "url": "http://localhost:3000"
      },
      {
        "url": "https://nodejs-hw-mongodb-1d4o.onrender.com"
      }
    ],
...
    "paths": {
      "/contacts": {
        "get": {
          "tags": [
            "Contacts"
          ],
          "summary": "Get all contacts",
          "operationId": "getAllContacts",
          "description": "Get all of your added contacts",
          "security": [
            {
              "bearerAuth": []
            }
          ],
          "parameters": [
            {
              "in": "query",
              "name": "page",
              "schema": {
                "type": "integer",
                "example": 2
              }
            },
            {
              "in": "query",
              "name": "perPage",
              "schema": {
                "type": "integer",
                "example": 30
              }
            },
            {
              "in": "query",
              "name": "sortBy",
              "schema": {
                "type": "string",
                "example": "name"
              },
              "description": "Possibility of sorting contacts by its characteristic"
            },
            {
              "in": "query",
              "name": "sortOrder",
              "schema": {
                "type": "string",
                "enum": [
                  "asc",
                  "desc"
                ],
                "example": "asc"
              }
            },
            {
              "in": "query",
              "name": "type",
              "schema": {
                "type": "string",
                "enum": [
                  "work",
                  "home",
                  "personal"
                ],
                "example": "work"
              }
            },
            {
              "in": "query",
              "name": "isFavourite",
              "schema": {
                "type": "boolean",
                "example": true
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Get all contacts",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "required": [
                      "status",
                      "message",
                      "data"
                    ],
                    "properties": {
                      "status": {
                        "type": "integer",
                        "example": 200
                      },
                      "message": {
                        "type": "string",
                        "example": "Successfully found contacts!"
                      },
                      "data": {
                        "type": "object",
                        "required": [
                          "data",
                          "page",
                          "perPage",
                          "totalItems",
                          "totalPages",
                          "hasNextPage",
                          "hasPreviousPage"
                        ],
                        "properties": {
                          "data": {
                            "type": "array",
                            "items": {
                              "$ref": "#/components/schemas/contact"
                            }
                          },
                          "page": {
                            "type": "number",
                            "example": 1
                          },
                          "perPage": {
                            "type": "number",
                            "example": 4
                          },
                          "totalItems": {
                            "type": "number",
                            "example": 8
                          },
                          "totalPages": {
                            "type": "number",
                            "example": 2
                          },
                          "hasNextPage": {
                            "type": "boolean",
                            "example": true
                          },
                          "hasPreviousPage": {
                            "type": "boolean",
                            "example": false
                          }
                        }
                      }
                    }
                  }
                }
              }
            },
            "401": {
              "$ref": "#/components/responses/401"
            }
          }
        },
        "post": {
          "tags": [
            "Contacts"
          ],
          "summary": "Add contact",
          "operationId": "addContact",
          "description": "Add contact",
          "security": [
            {
              "bearerAuth": []
            }
          ],
          "requestBody": {
            "content": {
              "multipart/form-data": {
                "schema": {
                  "type": "object",
                  "required": [
                    "name",
                    "phoneNumber",
                    "contactType"
                  ],
                  "properties": {
                    "name": {
                      "description": "Contact name",
                      "type": "string"
                    },
                    "phoneNumber": {
                      "description": "Contact phone number",
                      "type": "string"
                    },
                    "email": {
                      "description": "Contact email",
                      "type": "string"
                    },
                    "isFavourite": {
                      "description": "Contact is favourite",
                      "type": "boolean"
                    },
                    "contactType": {
                      "description": "Contact number type",
                      "type": "string",
                      "enum": [
                        "work",
                        "home",
                        "personal"
                      ]
                    },
                    "photo": {
                      "description": "Contact photo",
                      "type": "string",
                      "format": "binary"
                    }
                  }
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "Add contact",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "required": [
                      "status",
                      "message",
                      "data"
                    ],
                    "properties": {
                      "status": {
                        "type": "integer",
                        "example": 201
                      },
                      "message": {
                        "type": "string",
                        "example": "Successfully created a contact!"
                      },
                      "data": {
                        "type": "object",
                        "$ref": "#/components/schemas/contact"
                      }
                    }
                  }
                }
              }
            },
            "401": {
              "$ref": "#/components/responses/401"
            }
          }
        }
      },
      "/contacts/{contactId}": {
        "get": {
          "tags": [
            "Contacts"
          ],
          "summary": "Get contact by id",
          "operationId": "getContactById",
          "description": "Get contact by id",
          "security": [
            {
              "bearerAuth": []
            }
          ],
          "parameters": [
            {
              "in": "path",
              "name": "contactId",
              "required": true,
              "schema": {
                "type": "string",
                "example": "65ca67e7ae7f10c88b598300"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Get student by id",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "required": [
                      "status",
                      "message",
                      "data"
                    ],
                    "properties": {
                      "status": {
                        "type": "integer",
                        "example": 200
                      },
                      "message": {
                        "type": "string",
                        "example": "Successfully found student with id 65ca67e7ae7f10c88b598300!"
                      },
                      "data": {
                        "type": "object",
                        "$ref": "#/components/schemas/contact"
                      }
                    }
                  }
                }
              }
            },
            "404": {
              "$ref": "#/components/responses/404"
            }
          }
        },
        "patch": {
          "tags": [
            "Contacts"
          ],
          "summary": "Patch contact by id",
          "operationId": "updateContact",
          "description": "Update contact characteristics by id",
          "security": [
            {
              "bearerAuth": []
            }
          ],
          "parameters": [
            {
              "in": "path",
              "name": "contactId",
              "required": true,
              "schema": {
                "type": "string",
                "example": "65ca67e7ae7f10c88b598300"
              }
            }
          ],
          "requestBody": {
            "content": {
              "multipart/form-data": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "name": {
                      "description": "Contact name",
                      "type": "string"
                    },
                    "phoneNumber": {
                      "description": "Contact phone number",
                      "type": "string"
                    },
                    "email": {
                      "description": "Contact email",
                      "type": "string"
                    },
                    "isFavourite": {
                      "description": "Contact is favourite",
                      "type": "boolean"
                    },
                    "contactType": {
                      "description": "Contact type",
                      "type": "string",
                      "enum": [
                        "work",
                        "home",
                        "personal"
                      ]
                    },
                    "photo": {
                      "description": "Contact avatar",
                      "type": "string",
                      "format": "binary"
                    }
                  }
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "Patch contact",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "required": [
                      "status",
                      "message",
                      "data"
                    ],
                    "properties": {
                      "status": {
                        "type": "integer",
                        "example": 200
                      },
                      "message": {
                        "type": "string",
                        "example": "Successfully patched a contact!"
                      },
                      "data": {
                        "type": "object",
                        "$ref": "#/components/schemas/contact"
                      }
                    }
                  }
                }
              }
            },
            "404": {
              "$ref": "#/components/responses/404"
            }
          }
        },
        "delete": {
          "tags": [
            "Contacts"
          ],
          "summary": "Delete contact",
          "operationId": "deleteContact",
          "description": "Delete contact by id",
          "security": [
            {
              "bearerAuth": []
            }
          ],
          "parameters": [
            {
              "in": "path",
              "name": "contactId",
              "required": true,
              "schema": {
                "type": "string",
                "example": "65ca67e7ae7f10c88b598300"
              }
            }
          ],
          "responses": {
            "204": {
              "description": "Contact successfully deleted!"
            },
            "404": {
              "$ref": "#/components/responses/404"
            }
          }
        }
      },
...
    },
    "components": {
      "schemas": {
...
        "contact": {
          "type": "object",
          "required": [
            "_id",
            "name",
            "phoneNumber",
            "contactType",
            "userId"
          ],
          "properties": {
            "_id": {
              "description": "Contact id",
              "type": "string",
              "example": "65ca67e7ae7f10c88b598300"
            },
            "name": {
              "description": "Contact name",
              "type": "string",
              "example": "Mariia Ilko"
            },
            "phoneNumber": {
              "description": "Contact phone number",
              "type": "string",
              "example": "+380905938475"
            },
            "email": {
              "description": "Contact email",
              "type": "string",
              "example": "contact@gmail.com"
            },
            "isFavourite": {
              "description": "Contact is favourite",
              "type": "boolean",
              "example": true
            },
            "contactType": {
              "description": "Contact type",
              "type": "string",
              "enum": [
                "work",
                "home",
                "personal"
              ],
              "example": "work"
            },
            "userId": {
              "description": "Contact user id",
              "type": "string",
              "example": "65e3secdd253b30065d54fu8"
            },
            "photo": {
              "description": "Contact photo in cloud",
              "type": "string",
              "example": "https://res.cloudinary.com/image/image.jpg"
            }
          }
        },
...
      },
      "securitySchemes": {
        "bearerAuth": {
          "type": "http",
          "scheme": "bearer"
        }
      },
      "responses": {
        "401": {
          "description": "Unauthorized error",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "status",
                  "message",
                  "data"
                ],
                "properties": {
                  "status": {
                    "type": "integer",
                    "example": 401
                  },
                  "message": {
                    "type": "string",
                    "example": "UnauthorizedError"
                  },
                  "data": {
                    "type": "object",
                    "required": [
                      "message"
                    ],
                    "properties": {
                      "message": {
                        "type": "string",
                        "example": "Access token expired"
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "404": {
          "description": "NotFound Error",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "status",
                  "message",
                  "data"
                ],
                "properties": {
                  "status": {
                    "type": "integer",
                    "example": 404
                  },
                  "message": {
                    "type": "string",
                    "example": "NotFoundError"
                  },
                  "data": {
                    "type": "object",
                    "required": [
                      "message"
                    ],
                    "properties": {
                      "message": {
                        "type": "string",
                        "example": "Contact not found"
                      }
                    }
                  }
                }
              }
            }
          }
        },
...
      }
    }
  }